fix: 주량등록에서 "칭호" 필드 추가

+++ sulsul-domain/src/main/kotlin/will/of/d/sulsul/alcohol/drinkingLimit/domain/DrinkingLimitEntity.kt
@@ -18,7 +18,7 @@ data class DrinkingLimitEntity(
     val kakaoUserId: Long,
     val drinkType: String,
     val glass: Int,
-    val alcoholAmount: Double,
+    val alcoholAmount: Int,
 
     @CreatedDate
     val createdAt: LocalDateTime = LocalDateTime.now(),
